+++ src/methods/unban_chat_member.rs
@@ -2,15 +2,13 @@ use super::Method;
 use serde::Serialize;
 use typed_builder::TypedBuilder;
 
-/// Use this method to unban a previously kicked user in a supergroup or channel. The user will not
-/// return to the group or channel automatically, but will be able to join via link, etc. The bot
-/// must be an administrator for this to work. Returns `true` on success.
+/// Use this method to unban a previously kicked user in a supergroup or channel. The user will not return to the group or channel automatically, but will be able to join via link, etc. The bot must be an administrator for this to work. Returns True on success.
 #[derive(Debug, TypedBuilder, Serialize)]
 pub struct UnbanChatMember {
-    /// Unique identifier for the target chat.
+    /// Unique identifier for the target group or username of the target supergroup or channel (in the format @username)
     pub chat_id: i64,
-    /// Unique identifier of the target user.
-    pub user_id: i32,
+    /// Unique identifier of the target user
+    pub user_id: i64,
 }
 
 impl_method!(UnbanChatMember, bool, "unbanChatMember");
